# Optimize array swap operator for basic cases