--- a/modules/dists/BlockDist.chpl
+++ b/modules/dists/BlockDist.chpl
@@ -45,7 +45,7 @@ private use CommDiagnostics;
 private use ChapelLocks;
 private use ChapelDebugPrint;
 private use LayoutCS;
-import Reflection;
+
 public use SparseBlockDist;
 //
 // These flags are used to output debug information and run extra
@@ -1535,14 +1535,12 @@ where this.sparseLayoutType == unmanaged DefaultDist &&
   return true;
 }
 
+// Block1 <=> Block2 
 proc BlockArr.doiSwap(arr) {
   coforall (loc, localArr1, localArr2) in zip(Locales,this.locArr, arr.locArr) {
     localArr1.myElems.doiSwap(localArr2.myElems);
+    localArr1.locRAD <=> localArr2.locRAD;
   }
-  /*for i in 0..#this.locArr.size {
-    this.locArr[i].myElems.doiSwap(arr.locArr[i].myElems);
-    }*/
-  //this.locArr[0].myElems.doiSwap(arr.locArr[0].myElems);
 }
 
 private proc _doSimpleBlockTransfer(Dest, destDom, Src, srcDom) {