--- a/test/optimizations/arraySwap/twoDimensionalStridedSwap.chpl
+++ b/test/optimizations/arraySwap/twoDimensionalStridedSwap.chpl
@@ -1,0 +1,8 @@
+var d = {1..10, 3..17 by 7};
+var a: [d] int;
+var b: [d] int;
+a = 1;
+b = 2;
+a<=>b;
+writeln(a);
+writeln(b);