--- a/test/optimizations/arraySwap/twoDimensionalSwap.chpl
+++ b/test/optimizations/arraySwap/twoDimensionalSwap.chpl
@@ -1,0 +1,6 @@
+const D = {20..25, 30..35};
+var a: [D] int = 1;
+var b: [D] int = 2;
+a <=> b;
+writeln(a);
+writeln(b);