… test/optimizations/arraySwap/blockSwap.good
… test/optimizations/arraySwap/twoDimensionalStridedSwap.good
… test/optimizations/arraySwap/twoDimensionalSwap.good