This changes the `<=>` operator to swap the pointer for Block and Default Rectangular arrays, for some cases where it was currently doing a deep copy. In order to use this optimization, the arrays to be swapped must:
 - not be an array view
 - have the same domain
 - be of the same type
 - implement the `doiSwap()` method

If all of these conditions are not met, swaps will still use the old implementation.

Performance comparison for swap on two Block Dist arrays:
| Number of Elems | before changes (s) | after changes (s) |
| ----------- | -----: | ----: |
| 100,000,000    | 0.032 | 0.0027 |
| 1,000,000,000 | 0.31 | 0.0034 |
| 10,000,000,000   | 5.1 | 0.0032 |

(It is now constant time for both block and default rectangular)

- [x] paratest gasnet