@bradcray 2020-08-13
I believe #12959 could be closed now that this has been merged (?).

As a follow-on to this, I keep going back and forth on whether we could also swap in the DR<->DR case if the two domains (a) describe the same indices or (b) have the same shape.  Need to think about it more (maybe it's clearer to you guys).