@e-kayrakli 2020-08-13
I think this PR does (a) already? https://github.com/chapel-lang/chapel/pull/16209/files#diff-0d0e6549c1f4c9c666d4369507450911

(b) could be interesting, but there is probably some work to be done to adjust the domains' internals post swap.