@e-kayrakli 2020-08-13
And I think this is a perf test where we see its impact: https://chapel-lang.org/perf/chapcs/?startdate=2020/04/17&enddate=2020/08/13&graphs=parboilstencil3dexecutiontime