@bradcray 2020-08-13
The case I was interested in for (a) would look like:

```chapel
var a: [20..25, 30..35] int = 1;
var b: [20..25, 30..35] int = 2;
a <=> b;
writeln(a);
writeln(b);
```

(which might also be handled, though the PR's description didn't suggest that to me).

For (b), my quick think was "if each array keeps its same domain (as I believe it should... we're only swapping the elements' values, not the arrays' identities), then if the size/shape of the arrays are the same we could swap the pointers to their buffers (i.e., the _ddata fields) without worrying about anything else."  That is, 

```chapel
var a: [20..25, 30..35] int = 1;
var b: [21..26, 29..34] int = 2;
a <=> b;
writeln(a);
writeln(b);
```

should work by just swapping the DR's ddata fields.  My slow thing, though, was "wait Brad, is it really that easy...?" (though I haven't come up with a reason it wouldn't be yet).

Note that I think (a) only works for DR domains (for Block domains we'd need to make sure that their distributions were equivalent as well... which would be a good additional extension);  (b) only works for DR domains (and the distribution equivalence would be much harder to calculate.