@mcdobe100 2020-08-13
One questionable behavior I have observed while doing this optimization is illustrated below, and I am not sure if this is what would be expected, but it definitely is not what I would expect.

For this test:
```chapel
var a: [0..3, 0..2] int = 1;
var b: [0..2, 0..3] int = 2;
a <=> b;
writeln(a);
writeln(b);
```
This is the result:
```
2 2 2
2 2 2
2 2 2
0 0 0
1 1 1 2
1 1 1 2
1 1 1 2
```
Looking at the code, this makes sense why it is happening, but as a user, this is not what I would expect at all.

If instead we use the `a.size == b.size` check for using the optimized swap, this is the result of the test:
```
2 2 2
2 2 2
2 2 2
2 2 2
1 1 1 1
1 1 1 1
1 1 1 1
```
Which seems much more desirable to me, but it is a possible there is a reason that wouldn't be wanted if there is something that I am not aware of for why the current implementation was chosen.