@bradcray 2020-08-13
I agree that the master behavior is wrong, though I don't think the size check is what we want either.  We only support array assignment (or at least only _should_ support array assignment) if the shapes of the two arrays are the same, so "4x3" for example.  And I think this same principle should apply to array swaps, since it's essentially a two-way assignment.  I think there is a .shape() query on domains and/or arrays that should be helpful here?  If not, you could check that the sizes of each .dim are the same.

Given that (a) works, I'd be curious whether the following works (where "works" would be "swaps but without using the new optimization"):

```chapel
var D1 = {1..10, 1..10} dmapped Block({1..10, 1..10});
var D2 = {1..10, 1..10} dmapped Block({1..1000, 1..1000});
var a: [D1] int = 1;
var b: [D2] int = 2;
a <=> b;
writeln(a);
writeln(b);
```

This is an interesting case because the shapes of the two arrays match, but their domains aren't aligned due to the different arguments given to their distribution.

And then one other case for multi-locale execution would be:

```chapel
var D = {1..10, 1..10};
var A: [D] int = 1;
on Locales[numLocales-1] {
  var B: [D] int = 2;
  A <=> B;
}
```

where in this case the arrays have the identical domain, but they're not on the same locale (so the swap should work but not use the pointer-swap optimization).