@mcdobe100 2020-08-13
Adding in the `.shape` check makes the earlier (b) that you had mentioned use the optimized swap.

Both of the cases that you mentioned will try to use the optimized swap, and they both result in (different) errors currently on master. I think that this can be resolved by checking if all of the local subdomains are equal, but that seems a little heavyweight to me (since `localSubdomains()` returns an iter), that being said, I can't think of a better solution. Did you have another method in mind for checking that? 

One thing that I am thinking, which works for both of these cases, but I am not sure if it would work in all cases is just comparing the local subdomain of the arrays on the calling locale, which might work since we have already checked that the domains are equal?