@mcdobe100 2020-08-13
After thinking about this a little bit more, I think that it might be better to implement a `doiLegalSwap()` or similar method as Brad had recommended before starting this optimization, I think that it would make the logic a bit easier to understand and the code in ChapelArray to look a little bit cleaner.