@bradcray 2020-08-13
I'm not sure whether this idea made it into the PR, but one of the concepts in the original notes I sent out was to have a `doiLegalSwap` method (or maybe `doiLegalOptimizedSwap` is a better name?) on the arrays which would do any additional checking to see whether the pointer swap approach was legal for the given pair of arrays.  For example, DefaultRectangular would check to see that the two arrays are stored on the same locale and return true if they are, false otherwise.  I think for Block distributions, we should be able to do the check that the swap is legal without checking conditions on every target locale.  For example, a simple case would be "if the two arrays' domains have the same indices and share the same distribution, the swap is legal; if not, it's not."  A more permissive check would be "if the two domains have the same indices and the block distributions' bounding boxes and target locale sets are the same, it's legal, otherwise it's not" (I think there may be an existing dsiEquals() method or the like on distributions that could answer this type of question?).  We could go further and say that if each locale owned an array of the same shape it's legal under certain conditions, but to me, that seems like too much work / overhead to capture a case that I don't think is likely to come up in practice, so not worth the trouble and execution-time expense.