@e-kayrakli 2020-08-13
After a chat with Ben, I agree that we need a stricter checks (and I think I have introduced a similar bug in this release cycle because I always thought `==` on domains would check distributions as well https://github.com/chapel-lang/chapel/blob/master/modules/internal/ChapelAutoLocalAccess.chpl#L44)

Interface-wise, I find it a bit of an overkill to add a new method only to check whether we can do a fast swap or not. We have some `doi` methods returning true/false based on whether they can do the requested operations. See `doiBulkTransfer` for example.

So, if we follow that precedence (which is what I prefer), `doiSwap` would return true if the swap was done and false otherwise. If it returned false, we can do the forall approach we do today. What do you think about that?