@bradcray 2020-08-13
Oh, that'd be fine as well.